Created Task 2 of Python for File Extension Finder
In this program, we first prompt the user to enter a filename using the input() function and store the input in the variable filename.

Next, we use the rfind() method to find the last occurrence of the '.' character in the filename. If the '.' character is found, the rfind() method returns the index of its last occurrence; otherwise, it returns -1.

If the '.' character is found in the filename, we extract the extension by slicing the string from the index of '.' to the end of the string. We add 1 to the index to exclude the '.' character from the extension.

Finally, we use the print() function to display the extension of the file using an f-string. If the '.' character is not found in the filename, we assume that the file doesn't have an extension and display an appropriate message.